move getPipedStatus function
